remove SwapQuoteCalculator: unnecessary abstraction
